swift-bluebox/requirements.txt
flask>=0.10
werkzeug>=0.10
python-swiftclient>=2.3